Data test utility works - we can build-JSON-like structure in a macro, turn it to Expr, let it appear in test code and compare it against sth
