+++ hearth-tests/src/main/scala-2/hearth/testdata/DataFixtures.scala
@@ -1,0 +1,16 @@
+package hearth
+package testdata
+
+import scala.language.experimental.macros
+import scala.reflect.macros.blackbox
+
+class DataFixtures(val c: blackbox.Context) extends MacroCommonsScala2 with DataSupports with DataFixturesImpl {
+
+  // TODO: create macro annotation which would allow to do the following
+
+  def exampleImpl: c.Expr[Data] = example
+}
+object DataFixtures {
+
+  def example: Data = macro DataFixtures.exampleImpl
+}
