+++ hearth-tests/src/main/scala/hearth/testdata/package.scala
@@ -1,0 +1,16 @@
+package hearth
+
+package object testdata {
+
+  /** JSON-like data for usage in tests.
+    *
+    * Since macro can have only 1 result we can generate multiple things at once and return it as something easily
+    * diffable.
+    */
+  type Data = Data.Impl
+
+  type Diff = List[DiffEntry]
+  implicit final class DiffOps(private val diff: Diff) extends AnyVal {
+    def render: String = diff.map(_.render).mkString("\n")
+  }
+}
